Remove BEMification mixin
At the time, it seemed amazing to be able to nest 
stuff semantically and generate selector names for
elements and modifiers. But there’s no way you can
easily search for complex class names this way,
and it’s actually quicker to write a selector! So,
no more BEM name creating mixins!